+++ docs/changelog.md
@@ -9,6 +9,7 @@
 ### Fix bugs
 
 - Fixed error `Too many open files` in Unix-based systems when downloading manga in any `pdf` format
+- Fixed false owner list name when executing command `list:<user-id>`
 
 ### Improvements